// AppAngular/src/app/client/client.ts
 export class Client{

    id: number;
    photo: string;
    nom: string;
    prenom: string;
    nomSociete: string;
    adresse: string;
    ville: string;
    codePostal: string;
    numeroTel: string;
    email: string;
    dateContact: Date;
    etatCustomer: string;
    commentaire: string;

 }